requirements.txt
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
pinecone
google-generativeai
python-dotenv
tenacity
python-docx
pypdf
fastapi
uvicorn
pydantic
httpx
langchain
langchain-community
langchain-text-splitters
langchain-google-genai
simple-salesforce
beautifulsoup4
